+++ app/src/main/java/com/monke/monkeybook/model/content/DefaultModelImpl.java
@@ -104,20 +104,20 @@ public class DefaultModelImpl extends BaseModelImpl implements IStationBookModel
                         .searchBook(analyzeSearchUrl.getSearchPath(),
                                 analyzeSearchUrl.getQueryMap(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             } else if (url.contains("?")) {
                 return getRetrofitString(analyzeSearchUrl.getSearchUrl())
                         .create(IHttpGetApi.class)
                         .searchBook(analyzeSearchUrl.getSearchPath(),
                                 analyzeSearchUrl.getQueryMap(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             } else {
                 return getRetrofitString(analyzeSearchUrl.getSearchUrl())
                         .create(IHttpGetApi.class)
                         .getWebContent(analyzeSearchUrl.getSearchPath(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -153,20 +153,20 @@ public class DefaultModelImpl extends BaseModelImpl implements IStationBookModel
                         .searchBook(analyzeSearchUrl.getSearchPath(),
                                 analyzeSearchUrl.getQueryMap(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             } else if (bookSourceBean.getRuleSearchUrl().contains("?")) {
                 return getRetrofitString(analyzeSearchUrl.getSearchUrl())
                         .create(IHttpGetApi.class)
                         .searchBook(analyzeSearchUrl.getSearchPath(),
                                 analyzeSearchUrl.getQueryMap(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             } else {
                 return getRetrofitString(analyzeSearchUrl.getSearchUrl())
                         .create(IHttpGetApi.class)
                         .getWebContent(analyzeSearchUrl.getSearchPath(),
                                 headerMap)
-                        .flatMap(this::analyzeSearchBook);
+                        .flatMap(response -> new SearchBook().analyzeSearchBook(response, TAG, name, bookSourceBean));
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -175,50 +175,6 @@ public class DefaultModelImpl extends BaseModelImpl implements IStationBookModel
                 emitter.onComplete();
             });
         }
-    }
-
-    private Observable<List<SearchBookBean>> analyzeSearchBook(final Response<String> response) {
-        return Observable.create(e -> {
-            try {
-                String baseURI;
-                okhttp3.Response networkResponse = response.raw().networkResponse();
-                if (networkResponse != null && networkResponse.request() != null) {
-                    baseURI = networkResponse.request().url().toString();
-                } else {
-                    baseURI = response.raw().request().url().toString();
-                }
-                Document doc = Jsoup.parse(response.body());
-                Elements booksE = AnalyzeElement.getElements(doc, bookSourceBean.getRuleSearchList());
-                if (null != booksE && booksE.size() > 0) {
-                    List<SearchBookBean> books = new ArrayList<>();
-                    for (int i = 0; i < booksE.size(); i++) {
-                        SearchBookBean item = new SearchBookBean();
-                        item.setTag(TAG);
-                        item.setOrigin(name);
-                        AnalyzeElement analyzeElement = new AnalyzeElement(booksE.get(i), baseURI);
-                        item.setAuthor(FormatWebText.getAuthor(analyzeElement.getResult(bookSourceBean.getRuleSearchAuthor())));
-                        item.setKind(analyzeElement.getResult(bookSourceBean.getRuleSearchKind()));
-                        item.setLastChapter(analyzeElement.getResult(bookSourceBean.getRuleSearchLastChapter()));
-                        item.setName(analyzeElement.getResult(bookSourceBean.getRuleSearchName()));
-                        item.setNoteUrl(analyzeElement.getResult(bookSourceBean.getRuleSearchNoteUrl()));
-                        if (isEmpty(item.getNoteUrl())) {
-                            item.setNoteUrl(baseURI);
-                        }
-                        item.setCoverUrl(analyzeElement.getResult(bookSourceBean.getRuleSearchCoverUrl()));
-                        if (!isEmpty(item.getName())) {
-                            books.add(item);
-                        }
-                    }
-                    e.onNext(books);
-                } else {
-                    e.onNext(new ArrayList<>());
-                }
-            } catch (Exception ex) {
-                ex.printStackTrace();
-                e.onNext(new ArrayList<>());
-            }
-            e.onComplete();
-        });
     }
 
     /**
